+++ tests/java_integers.rs
@@ -1,6 +1,6 @@
 #![cfg(feature = "invocation")]
 
-use jni::{errors::Result, jni_sig, objects::JValue};
+use jni::{errors::Result, jni_sig, jni_str, objects::JValue};
 
 mod util;
 use util::{attach_current_thread, print_exception};
@@ -13,18 +13,21 @@ fn test_java_integers() {
         for value in -10..10 {
             env.with_local_frame(16, |env| -> Result<()> {
                 let integer_value = env.new_object(
-                    c"java/lang/Integer",
+                    jni_str!("java/lang/Integer"),
                     jni_sig!("(I)V"),
                     &[JValue::Int(value)],
                 )?;
 
-                let values_array =
-                    env.new_object_array(array_length, c"java/lang/Integer", &integer_value)?;
+                let values_array = env.new_object_array(
+                    array_length,
+                    jni_str!("java/lang/Integer"),
+                    &integer_value,
+                )?;
 
                 let result = env
                     .call_static_method(
-                        c"java/util/Arrays",
-                        c"binarySearch",
+                        jni_str!("java/util/Arrays"),
+                        jni_str!("binarySearch"),
                         jni_sig!("([Ljava/lang/Object;Ljava/lang/Object;)I"),
                         &[
                             JValue::Object(&values_array),