+++ tests/ui/fail/const-jnistr-from-cstr.rs
@@ -1,5 +1,5 @@
 use jni::strings::JNIStr;
 
 fn main() {
-    const INVALID_MUTF8_CSTR: &JNIStr = JNIStr::from_cstr(c"invalid mutf8: 🦀");
+    const INVALID_MUTF8_CSTR: &JNIStr = JNIStr::from_cstr(c"invalid mutf8: 🦀").unwrap();
 }