+++ tests/util/mod.rs
@@ -1,7 +1,8 @@
 use std::sync::{Arc, Once};
 
 use jni::{
-    errors::Result, jni_sig, objects::JValue, sys::jint, Env, InitArgsBuilder, JNIVersion, JavaVM,
+    errors::Result, jni_sig, jni_str, objects::JValue, sys::jint, Env, InitArgsBuilder, JNIVersion,
+    JavaVM,
 };
 
 mod example_proxy;
@@ -36,8 +37,8 @@ pub fn jvm() -> &'static Arc<JavaVM> {
 #[allow(dead_code)]
 pub fn call_java_abs(env: &mut Env, value: i32) -> i32 {
     env.call_static_method(
-        c"java/lang/Math",
-        c"abs",
+        jni_str!("java/lang/Math"),
+        jni_str!("abs"),
         jni_sig!("(I)I"),
         &[JValue::from(value as jint)],
     )